Update: modularization and fix zooming issues
